+++ src/streamdiffusion/acceleration/tensorrt/runtime_engines/engine_pool.py
@@ -10,9 +10,9 @@ import torch
 from polygraphy import cuda
 
 from .controlnet_engine import ControlNetModelEngine, HybridControlNet
-from .models.controlnet_models import create_controlnet_model
-from .builder import EngineBuilder, create_onnx_path
-from ...model_detection import detect_model
+from ..models.controlnet_models import create_controlnet_model
+from ..builder import EngineBuilder, create_onnx_path
+from ....model_detection import detect_model
 
 # Set up logger for this module
 logger = logging.getLogger(__name__)
